refactor: NavigationBar 파일 분리
[#804]
